+++ nexus/logs/api/routers.py
@@ -5,7 +5,8 @@ from nexus.logs.api.views import (
     MessageHistoryViewset,
     TagPercentageViewSet,
     MessageDetailViewSet,
-    ConversationContextViewset
+    ConversationContextViewset,
+    InlineConversationsViewset
 )
 
 
@@ -17,6 +18,7 @@ urlpatterns = [
     path('<project_uuid>/tags-analytics/', TagPercentageViewSet.as_view({'get': 'list'}), name='list-tag-percentage'),
     path('<project_uuid>/message-detail/<log_id>/', MessageDetailViewSet.as_view(), name="message-detail"),
     path('<project_uuid>/conversation-context/', ConversationContextViewset.as_view({'get': 'list'}), name='list-conversation-context'),
+    path('<project_uuid>/conversations/', InlineConversationsViewset.as_view({'get': 'list'}), name='list-inline-conversations'),
 ]
 
 urlpatterns.append(path('prometheus/', include('django_prometheus.urls')))