+++ nexus/projects/models.py
@@ -44,6 +44,8 @@ class Project(BaseModel, SoftDeleteModel):
 
     human_support = models.BooleanField(default=False)
     human_support_prompt = models.TextField(null=True, blank=True)
+    rationale_switch = models.BooleanField(default=False)
+    inline_agent_switch = models.BooleanField(default=False)
 
     def __str__(self):
         return f'{self.uuid} - Project: {self.name} - Org: {self.org.name}'