+++ nexus/task_managers/file_database/bedrock.py
@@ -1,8 +1,8 @@
+import os
 import uuid
 import json
 import time
 from typing import TYPE_CHECKING
-
 from io import BytesIO
 
 from dataclasses import dataclass
@@ -23,7 +23,7 @@ from nexus.task_managers.file_database.file_database import FileDataBase, FileRe
 
 from nexus.agents.models import Agent, Credential, Team
 
-from nexus.agents.components import get_all_formats, get_all_formats_list
+from nexus.agents.components import get_all_formats_list
 
 if TYPE_CHECKING:
     from router.entities import Message
@@ -579,11 +579,14 @@ class BedrockFileDatabase(FileDataBase):
 
         sessionState["sessionAttributes"] = {"credentials": json.dumps(credentials, default=str)}
 
+        time_now = pendulum.now("America/Sao_Paulo")
+        llm_formatted_time = f"Today is {time_now.format('dddd, MMMM D, YYYY [at] HH:mm:ss z')}"
+
         sessionState["promptSessionAttributes"] = {
             # "format_components": get_all_formats(),
             "contact_urn": message.contact_urn,
             "contact_fields": message.contact_fields_as_json,
-            "date_time_now": pendulum.now("America/Sao_Paulo").isoformat(),
+            "date_time_now": llm_formatted_time,
             "project_id": message.project_uuid,
             "specific_personality": json.dumps({
                 "occupation": agent.role,
@@ -600,13 +603,13 @@ class BedrockFileDatabase(FileDataBase):
             })
 
         if team.human_support:
-            sessionState["promptSessionAttributes"] = {
+            sessionState["promptSessionAttributes"].update({
                 "human_support": json.dumps({
                     "project_id": message.project_uuid,
                     "contact_id": message.contact_urn,
                     "business_rules": team.human_support_prompt
                 })
-            }
+            })
 
         try:
             response = self.bedrock_agent_runtime.invoke_agent(
@@ -1073,6 +1076,11 @@ class BedrockFileDatabase(FileDataBase):
 
         return _lambda_iam_role["Role"]["Arn"]
 
+    def upload_inline_traces(self, data, key):
+        custom_bucket = os.getenv('AWS_BEDROCK_INLINE_TRACES_BUCKET')
+        bytes_stream = BytesIO(data.encode('utf-8'))
+        self.s3_client.upload_fileobj(bytes_stream, custom_bucket, key)
+
     def upload_traces(self, data, key):
         bytes_stream = BytesIO(data.encode('utf-8'))
         self.s3_client.upload_fileobj(bytes_stream, self.bucket_name, key)
@@ -1080,6 +1088,14 @@ class BedrockFileDatabase(FileDataBase):
     def get_trace_file(self, key):
         try:
             response = self.s3_client.get_object(Bucket=self.bucket_name, Key=key)
+            return response['Body'].read().decode('utf-8')
+        except self.s3_client.exceptions.NoSuchKey:
+            return []
+
+    def get_inline_trace_file(self, key):
+        try:
+            custom_bucket = os.getenv('AWS_BEDROCK_INLINE_TRACES_BUCKET')
+            response = self.s3_client.get_object(Bucket=custom_bucket, Key=key)
             return response['Body'].read().decode('utf-8')
         except self.s3_client.exceptions.NoSuchKey:
             return []